+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+Runtime Behavior Changes
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+* finagle: Bump version of Jackson to 2.13.1. ``PHAB_ID=D808049``
+
 21.12.0
 -------
 
@@ -79,6 +84,9 @@ Breaking API Changes
 
 * finagle-core: `c.t.f.loadbalancer.distributor.AddressedFactory` has been removed. Use
 `c.t.f.loadbalancer.EndpointFactory` directly instead. ``PHAB_ID=D751145``
+
+Runtime Behavior Changes
+~~~~~~~~~~~~~~~~~~~~~~~~
 
 * finagle-core: Moved `c.t.finagle.stats.LoadedStatsReceiver` and `c.t.finagle.stats.DefaultStatsReceiver`
   from the finagle-core module to util-stats.  ``PHAB_ID=D763497``